[ProcessModules] Updated ProcessModules to easily remove redundant code and make adding new robots easier
