+++ src/pycram/process_modules/hsrb_process_modules.py
@@ -2,7 +2,7 @@ import numpy as np
 from threading import Lock
 from typing_extensions import Any
 
-from .default_process_modules import DefaultDetecting, DefaultDetectingReal
+from .default_process_modules import *
 from ..datastructures.enums import ExecutionType
 from ..external_interfaces.tmc import tmc_gripper_control, tmc_talk
 from ..robot_description import RobotDescription
@@ -20,24 +20,6 @@ import io
 from ..ros.logging import logdebug
 
 
-class HSRBNavigation(ProcessModule):
-    """
-    The process module to move the robot from one position to another.
-    """
-
-    def _execute(self, desig: MoveMotion):
-        robot = World.robot
-        robot.set_pose(desig.target)
-
-
-class HSRBDetecting(ProcessModule):
-    """
-    This process module tries to detect an object with the given type. To be detected the object has to be in
-    the field of view of the robot.
-    """
-   # pass
-    def _execute(self, desig: DetectingMotion) -> Any:
-        pass
 
 
 ###########################################################
@@ -65,16 +47,6 @@ class HSRBMoveHeadReal(ProcessModule):
     def _execute(self, desig: LookingMotion):
         target = desig.target
         giskard.move_head_to_pose(target)
-
-
-class HSRBDetectingReal(ProcessModule):
-    """
-    Process Module for the real HSRB that tries to detect an object fitting the given object description. Uses Robokudo
-    for perception of the environment.
-    """
-
-    def _execute(self, desig: DetectingMotion) -> Any:
-        pass
 
 
 class HSRBMoveTCPReal(ProcessModule):
@@ -200,27 +172,17 @@ class HSRBTalkSemiReal(ProcessModule):
 ###########################################################
 ########## HSRB MANAGER ###############
 ###########################################################
-class HSRBManager(ProcessModuleManager):
+class HSRBManager(DefaultManager):
 
     def __init__(self):
-        super().__init__("hsrb")
+        super().__init__()
+        self.robot_name = "hsrb"
         self._navigate_lock = Lock()
-        self._pick_up_lock = Lock()
-        self._place_lock = Lock()
-        self._looking_lock = Lock()
-        self._detecting_lock = Lock()
-        self._move_tcp_lock = Lock()
-        self._move_arm_joints_lock = Lock()
-        self._world_state_detecting_lock = Lock()
-        self._move_joints_lock = Lock()
-        self._move_gripper_lock = Lock()
-        self._open_lock = Lock()
-        self._close_lock = Lock()
-        self._talk_lock = Lock()
+
 
     def navigate(self):
         if ProcessModuleManager.execution_type == ExecutionType.SIMULATED:
-            return HSRBNavigation(self._navigate_lock)
+            return DefaultNavigation(self._navigate_lock)
         elif ProcessModuleManager.execution_type == ExecutionType.REAL:
             return HSRBNavigationReal(self._navigate_lock)
         elif ProcessModuleManager.execution_type == ExecutionType.SEMI_REAL:
